Upgraded VS project to 2022

+++ tutorial29_youtube/skinned_mesh.h
@@ -44,7 +44,7 @@ public:
 
     void Render();
 
-    uint NumBones() const
+    size_t NumBones() const
     {
         return m_BoneNameToIndexMap.size();
     }
